Remove unused functions in kubernetes package
